+++ src/cpu/chip8/mod.rs
@@ -5,11 +5,14 @@ use parcel::Parser;
 
 use super::ExecuteMut;
 
+mod display;
 mod memory;
 mod microcode;
 mod operations;
 mod register;
 mod u12;
+
+use display::Display;
 
 /// Defines a trait for implementing random number generation.
 pub trait GenerateRandom<T> {
@@ -70,68 +73,6 @@ pub enum KeyInputValue {
     KeyD,
     KeyE,
     KeyF,
-}
-
-/// Display mimics the display matrix for the CHIP-8 isa.
-#[derive(Debug, Clone, Copy)]
-pub struct Display {
-    inner: [[bool; 64]; 32],
-}
-
-impl Default for Display {
-    fn default() -> Self {
-        Self {
-            inner: [[false; 64]; 32],
-        }
-    }
-}
-
-impl Display {
-    /// Returns the maximum number of columns per row.
-    pub fn x_max() -> usize {
-        64
-    }
-
-    /// Returns the maximum number of rows.
-    pub fn y_max() -> usize {
-        32
-    }
-
-    /// gets the value of the pixel specified by the cartesian coordinates `x`,
-    /// `y`. If the coordinates are within range, an `Option::Some(bool)` with
-    /// the value of the pixel is returned. Otherwise `Option::None` is
-    /// returned.
-    pub fn pixel(&self, x: usize, y: usize) -> Option<bool> {
-        if x < Self::x_max() && y < Self::y_max() {
-            Some(self.inner[y][x])
-        } else {
-            None
-        }
-    }
-
-    /// sets the value of the pixel specified by the cartesian coordinates `x`,
-    /// `y` to the boolean value specified by `pixel_on`. If the coordinates
-    /// are within range, a `(Self, Option::Some(bool))` with the previous
-    /// value of the pixel is returned. Otherwise `(Self, Option::None)` is
-    /// returned.
-    pub fn write_pixel(mut self, x: usize, y: usize, pixel_on: bool) -> (Self, Option<bool>) {
-        let modified = self.write_pixel_mut(x, y, pixel_on);
-        (self, modified)
-    }
-
-    /// sets the value of the pixel specified by the cartesian coordinates `x`,
-    /// `y` to the boolean value specified by `pixel_on`. If the coordinates
-    /// are within range, an `Option::Some(bool)` with the previous value of
-    /// the pixel is returned. Otherwise `Option::None` is returned.
-    pub fn write_pixel_mut(&mut self, x: usize, y: usize, pixel_on: bool) -> Option<bool> {
-        if let Some(previous_value) = self.pixel(x, y) {
-            // if we can get the previous value, it's safe to write a new one.
-            self.inner[y][x] = pixel_on;
-            Some(previous_value)
-        } else {
-            None
-        }
-    }
 }
 
 /// Represents the address the program counter is set to on chip reset.
@@ -558,7 +499,7 @@ impl<R> crate::cpu::ExecuteMut<microcode::SetDisplayRange> for Chip8<R> {
         // add 1 to cover non-inclusive range.
         let modified = ((end_y - start_y) * Display::x_max()) + (end_x - start_x) + 1;
         self.display
-            .inner
+            .unwrap()
             .iter_mut()
             .flatten()
             .skip(start_offset)
@@ -632,7 +573,7 @@ mod tests {
             &microcode::SetDisplayRange::new((0, 0), (63, 31), true),
         );
 
-        for (pos, &v) in cpu.display.inner.iter().flatten().enumerate() {
+        for (pos, &v) in cpu.display.unwrap().iter().flatten().enumerate() {
             assert_eq!(true, v, "no match for position: {}, {}", pos, v);
         }
     }
